requirements.txt
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
chatgptpy==1.0.5
click==8.1.3
colorama==0.4.6
cssselect2==0.7.0
gTTS==2.3.0
idna==3.4
install==1.3.5
lxml==4.9.1
Pillow==9.3.0
PyAudio==0.2.12
pyChatGPT==0.1.0
python-dotenv==0.21.0
reportlab==3.6.12
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
SpeechRecognition==3.9.0
svglib==1.4.1
tinycss2==1.2.1
tls-client==0.1.5
urllib3==1.26.13
uuid==1.30
webencodings==0.5.1
